Completed user journey for create/voting on talks.
The application is getting close to completion, most of the logical work
is done on the frontend. Completed the talksSlice of the redux store with
applicable reducers for voting, as well as updating the user's votes.
Fixed the voting logic on the backend to make it all tick along with
adding a proper global exception guard for detail responses to the frontend.